feat: additional test and logcollector arch separation
