+++ src/modules/logcollector/src/logcollector.cpp
@@ -73,7 +73,7 @@ void Logcollector::Setup(std::shared_ptr<const configuration::ConfigurationParse
 
     SetupFileReader(configurationParser);
 
-    AddPlatformSpecificReader(configurationParser, *this);
+    AddPlatformSpecificReader(configurationParser);
 }
 
 void Logcollector::SetupFileReader(const std::shared_ptr<const configuration::ConfigurationParser> configurationParser)
