+++ src/modules/logcollector/src/logcollector_unix.cpp
@@ -1,0 +1,17 @@
+#include <config.h>
+#include <timeHelper.h>
+
+#include <chrono>
+#include <map>
+
+#include <logcollector.hpp>
+
+namespace logcollector
+{
+
+void Logcollector::AddPlatformSpecificReader([[maybe_unused]] std::shared_ptr<const configuration::ConfigurationParser> configurationParser)
+{
+
+}
+
+}
