+++ src/modules/logcollector/tests/unit/event_reader_mocks.hpp
@@ -38,3 +38,25 @@ private:
 
 };
 
+class LogcollectorMock : public Logcollector {
+public:
+    LogcollectorMock() {
+        ON_CALL(*this, AddReader(::testing::_))
+            .WillByDefault(::testing::Invoke([this](std::shared_ptr<IReader> reader) {
+                this->Logcollector::AddReader(reader);
+            })
+        );
+    }
+
+    MOCK_METHOD(void, AddReader, (std::shared_ptr<IReader> reader), (override));
+    MOCK_METHOD(void, EnqueueTask, (Awaitable task), (override));
+    MOCK_METHOD(void, SendMessage, (const std::string& channel, const std::string& message,  const std::string& collectorType), (override));
+    boost::asio::awaitable<void> Wait([[maybe_unused]]std::chrono::milliseconds ms)
+    {
+        return Logcollector::Wait(ms);
+    }
+    void Stop()
+    {
+        Logcollector::Stop();
+    }
+};
