+++ src/modules/logcollector/tests/unit/logcollector_mock.hpp
@@ -22,16 +22,6 @@ public:
     }
     MOCK_METHOD(void, AddReader, (std::shared_ptr<IReader> reader), (override));
     MOCK_METHOD(void, EnqueueTask, (Awaitable task), (override));
-    // TODO:
-    // MOCK_METHOD(void, SendMessage, (const std::string& channel, const std::string& message, int collectorType), ());
-    boost::asio::awaitable<void> Wait([[maybe_unused]]std::chrono::milliseconds ms)
-    {
-        return Logcollector::Wait(ms);
-    }
-    void Stop()
-    {
-        Logcollector::Stop();
-    }
 };
 
 class PushMessageMock {
